pyproject.toml
[tool.poetry]
name = "iamport"
version = "0.1.0"
description = ""
authors = ["kimdooh <johnkdo2020@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "<3.0"
djangorestframework = "^3.12.2"
django-extensions = "^3.1.0"
notebook = "^6.2.0"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
cfn-flip = "^1.2.3"
argon2-cffi = "^20.1.0"
supervisor = "^4.2.2"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
